pyproject.toml
[project]
name = "tacthru"
version = "1.0.0"
requires-python = ">=3.10"

dependencies = [
    "opencv-python>=4.11.0.86",
    "numpy>=2.2.6",
    "scikit-learn>=1.7.0",
    "pandas>=2.3.3",
    "loguru>=0.7.3",
    "pre-commit>=4.3.0",
]


[project.optional-dependencies]
umi = [
    "accelerate>=1.9.0",
    "atomics>=1.0.3",
    "av>=15.0.0",
    "dearpygui>=2.1.0",
    "diffusers>=0.34.0",
    "dill>=0.4.0",
    "einops>=0.8.1",
    "gym>=0.26.2",
    "hydra-core>=1.3.2",
    "imagecodecs>=2025.3.30",
    "matplotlib>=3.10.3",
    "meshcat>=0.3.2",
    "natsort>=8.4.0",
    "numcodecs<=0.16",
    "nvidia-ml-py>=13.580.82",
    "omegaconf>=2.3.0",
    "plotly>=6.2.0",
    "pyarrow>=20.0.0",
    "rich>=14.1.0",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "timm>=1.0.17",
    "torch==2.7.0+cu128",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "transformers>=4.55.4",
    "wandb>=0.21.0",
    "webdataset>=1.0.2",
    "zarr==2.16.1",
]

[tool.ruff]
line-length = 120
select = ["F", "E", "W", "UP"]
ignore = ["F401", "E402", "F541", "UP006", "UP035", "I001", "I002"]

[tool.uv]
index-strategy = "unsafe-best-match"
index-url = "https://download.pytorch.org/whl/cu128"
extra-index-url = ["https://pypi.org/simple"]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
